+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### GA Release v1.0.0 (February 17, 2018)
 
 * Updated sources
-  * Added [https://github.com/StevenBlack/hosts/](https://raw.githubusercontent.com/StevenBlack/hosts/master/alternates/fakenews-gambling-porn-social/hosts)
-  * Removed Yoyo, since the new sources make it redundant
+  * Added hosts source [https://github.com/StevenBlack/hosts/](https://raw.githubusercontent.com/StevenBlack/hosts/master/alternates/fakenews-gambling-porn-social/hosts)
+  * Removed Yoyo and raw.github.com, since the new sources make them redundant
 * Fixed bug where "domains" was reported as "hosts" source type
 
 ### Release Candidate v1.0.0.rc5 (February 4, 2018)